refactor: move code into proper directory
