+++ front/lib/api/llm/utils/openai_like/responses/conversation_to_openai.ts
