+++ front/lib/api/llm/utils/openai_like/responses/openai_to_events.ts
