+++ front/lib/api/llm/utils/openai_like/responses/test/fixtures/llm_events/reasoning.ts
