+++ front/lib/api/llm/utils/openai_like/responses/test/fixtures/model_output/function_call.ts
