+++ front/lib/api/llm/utils/openai_like/responses/test/openai_to_events.test.ts
@@ -1,11 +1,11 @@
 import { describe, expect, it } from "vitest";
 
 import { createAsyncGenerator } from "@app/lib/api/llm/utils";
-import * as openai_to_events from "@app/lib/api/llm/utils/openai_to_events";
-import { functionCallLLMEvents } from "@app/lib/api/llm/utils/test/fixtures/llm_events/function_call";
-import { reasoningLLMEvents } from "@app/lib/api/llm/utils/test/fixtures/llm_events/reasoning";
-import { functionCallModelEvents } from "@app/lib/api/llm/utils/test/fixtures/model_output/function_call";
-import { reasoningModelOutput } from "@app/lib/api/llm/utils/test/fixtures/model_output/reasoning";
+import * as openai_to_events from "@app/lib/api/llm/utils/openai_like/responses/openai_to_events";
+import { functionCallLLMEvents } from "@app/lib/api/llm/utils/openai_like/responses/test/fixtures/llm_events/function_call";
+import { reasoningLLMEvents } from "@app/lib/api/llm/utils/openai_like/responses/test/fixtures/llm_events/reasoning";
+import { functionCallModelEvents } from "@app/lib/api/llm/utils/openai_like/responses/test/fixtures/model_output/function_call";
+import { reasoningModelOutput } from "@app/lib/api/llm/utils/openai_like/responses/test/fixtures/model_output/reasoning";
 
 const metadata = {
   clientId: "openai_responses",
